(API) Minor refactor for project create with github
