+++ src-tauri/src/projects/mod.rs
@@ -1,5 +1,8 @@
 use std::{collections::HashMap, path::Path};
 
+use crate::github::save_deploy_key_to_repository;
+
+use super::github::clone_github_repository;
 use git2::{build::RepoBuilder, Cred, RemoteCallbacks};
 use reqwest::{header, StatusCode};
 use ssh_key::{rand_core::OsRng, Algorithm, LineEnding, PrivateKey};
@@ -20,10 +23,6 @@ fn get_repository_owner_and_name(github_repo_url: &str) -> Result<(&str, &str), 
     let name = split.next().unwrap();
     let owner = split.next().unwrap();
     Ok((owner, name))
-}
-
-fn get_repository_ssh_url(owner: &str, name: &str) -> String {
-    format!("git@github.com:{}/{}.git", owner, name)
 }
 
 #[tauri::command]
@@ -47,76 +46,21 @@ pub async fn create_project(
             LINE_ENDING,
         )
         .unwrap();
-    let ssh_public_key = ssh_private_key.public_key().clone().to_string();
 
-    // We save the SSH public key to GitHub using the API
-    // https://docs.github.com/en/rest/deploy-keys/deploy-keys?apiVersion=2022-11-28
-    let mut payload = HashMap::new();
-    payload.insert("title", "opshala-desktop-app");
-    payload.insert("key", ssh_public_key.as_str());
-    payload.insert("read_only", "false");
-
-    let mut headers = reqwest::header::HeaderMap::new();
-    headers.insert(
-        header::ACCEPT,
-        "application/vnd.github+json".parse().unwrap(),
-    );
-    headers.insert(
-        header::AUTHORIZATION,
-        format!("Bearer {}", github_token).parse().unwrap(),
-    );
-    headers.insert("X-GitHub-Api-Version", "2022-11-28".parse().unwrap());
-    headers.insert(header::USER_AGENT, "opshala-desktop-app".parse().unwrap());
-    let client = reqwest::Client::new();
-    let response = client
-        .post("https://api.github.com/repos/brainless/opshala-test/keys")
-        .json(&payload)
-        .headers(headers)
-        .send()
+    save_deploy_key_to_repository(owner, name, ssh_private_key.public_key(), github_token)
         .await
         .unwrap();
 
-    if response.status() != StatusCode::CREATED {
-        println!("{:?}", response.headers());
-        return Err("Error while creating deploy key on GitHub".to_owned());
-    }
-
-    // We clone the repository using the SSH URL and the SSH private key
-    // Code take from https://docs.rs/git2/latest/git2/build/struct.RepoBuilder.html#example
-    let mut callbacks = RemoteCallbacks::new();
-    callbacks.credentials(|_url, username_from_url, _allowed_types| {
-        Cred::ssh_key(
-            username_from_url.unwrap(),
-            None,
-            std::path::Path::new(
-                app_handle
-                    .path_resolver()
-                    .app_data_dir()
-                    .unwrap()
-                    .join(format!("ssh-key-{}-{}", owner, name))
-                    .as_path(),
-            ),
-            None,
-        )
-    });
-
-    // Prepare fetch options
-    let mut fetch_options = git2::FetchOptions::new();
-    fetch_options.remote_callbacks(callbacks);
-
-    // Prepare builder
-    let mut builder = RepoBuilder::new();
-    builder.fetch_options(fetch_options);
-
-    // Clone the project
-    builder
-        .clone(
-            &get_repository_ssh_url(owner, name),
-            Path::new(parent_folder_path)
-                .join(Path::new(name))
-                .as_path(),
-        )
-        .unwrap();
+    clone_github_repository(
+        owner,
+        name,
+        parent_folder_path,
+        app_handle
+            .path_resolver()
+            .app_data_dir()
+            .unwrap()
+            .join(format!("ssh-key-{}-{}", owner, name)),
+    );
     Ok(())
 }
